Replace 'canGoHome' with 'showRestart' across the app
The purpose of this commit is to update the boolean attribute 'canGoHome' with 'showRestart' in multiple dart files for improved clarity. A 'menu' option has also been added in the 'MenuIconType' in the default_layout.dart file to extend the app's navigation. The 'sheet' package was added in the dependencies for the 'vyuh_feature_system' package.