Refactor feature init and add dispose function
The code has been refactored to change FeatureInitFunction to VoidFutureFunction and to add a new dispose function in the FeatureDescriptor. Additionally, an initFeatures function has been added to run a cleanup before initializing new features. The feature list has been changed to a FeatureBuilder. The new implementation ensures features are preloaded, initialized, and disposed properly.