Refactor code in vyuh_feature_developer for better readability and organization
The commit transfers 'FeatureHeroCard' component from components directory to feature_detail.dart file directly, enhancing code simplicity and maintainability. On top of this, the pubspec.lock file has also been updated to reflect the latest versions of dependencies resulting from the restructuring. The changes aim for increased efficiency and better readability and organization of the code.